[refactor] novel, episode exception 하위 errorcode 패키지 생성 및 ErrorCode 이동 - #12
