+++ novelservice/src/main/java/com/iucyh/novelservice/novel/exception/InvalidNovelCursor.java
@@ -1,6 +1,7 @@
 package com.iucyh.novelservice.novel.exception;
 
 import com.iucyh.novelservice.common.exception.ServiceException;
+import com.iucyh.novelservice.novel.exception.errorcode.NovelErrorCode;
 
 public class InvalidNovelCursor extends ServiceException {
